+++ src/rust/src/types.rs
@@ -45,6 +45,8 @@ pub static DEPRECATED_IN_41: LazyPyImport =
     LazyPyImport::new("cryptography.utils", &["DeprecatedIn41"]);
 pub static DEPRECATED_IN_42: LazyPyImport =
     LazyPyImport::new("cryptography.utils", &["DeprecatedIn42"]);
+pub static DEPRECATED_IN_43: LazyPyImport =
+    LazyPyImport::new("cryptography.utils", &["DeprecatedIn43"]);
 
 pub static ENCODING: LazyPyImport = LazyPyImport::new(
     "cryptography.hazmat.primitives.serialization",